refactor: move py scripts to folder "alignment_drift"
